refactor(markdown): move h2mSync to h2m module
