+++ markdown/index.ts
@@ -1,2 +1,3 @@
 export * from "./utils/index.js";
 export * from "./m2h/index.js";
+export * from "./h2m/index.js";
